Changed priority to bigint in DB

+++ packages/persistance/prisma/migrations/20260117134313_transaction_priority/migration.sql
@@ -1,7 +1,7 @@
 -- CreateTable
 CREATE TABLE "TransactionPriority" (
     "transactionHash" TEXT NOT NULL,
-    "priority" INTEGER NOT NULL,
+    "priority" BIGINT NOT NULL,
 
     CONSTRAINT "TransactionPriority_pkey" PRIMARY KEY ("transactionHash")
 );
